Use semi-rigid patch matching
This approach first calculates an optimal multiplier that gives the best
global match between the reference image and the working image and then uses
this result as the multiplier when estimating patch shifts. This should
improve shift estimates in smooth regions as it does not allow the patches
to match with an arbitrary multiplier.